Refactor move simple functions to a separate file
